Change to not using "dot config" approach.
Originally used config folders prefixed with a period. This seemed like a good ideal, as it would, in most cases, cause those folders to be hidden from the user. But this has negatives, too. Since these folders are not entirely managed by the system -- and require user editing -- this was probably the wrong way to go. Also, "dot" folders are hidden in Eclipse by default.